apitodo/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.10
django-ckeditor==6.1.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-stdimage==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
hubspot-api-client==4.0.6
idna==3.3
mysql-connector-python==8.0.27
Pillow==8.1.2
progressbar2==3.53.1
psycopg2==2.8.6
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.2.0
